fix comment
-move favMovies functions from user.service to fav-movies.service
-fix unnecessary
