+++ src/fav-movies/fav-movies.service.ts
@@ -1,16 +1,30 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  InternalServerErrorException,
+} from '@nestjs/common';
 import { HttpService } from '@nestjs/axios';
-import { map } from 'rxjs/operators';
 import { Observable } from 'rxjs';
 import { AxiosResponse } from 'axios';
 import { FavMoviesDto } from './dto/fav-movies.dto';
 import { API_URL } from 'src/constant/constant';
+import { UserEntity } from 'src/user/user.entity';
+import { InjectRepository } from '@nestjs/typeorm';
+import { Repository } from 'typeorm';
+import { FavMoviesEntity } from './fav-movies.entity';
+import { lastValueFrom, map } from 'rxjs';
+import { UserService } from 'src/user/user.service';
 
 @Injectable()
 export class FavMoviesService {
   private apiUrl = API_URL;
 
-  constructor(private http: HttpService) {}
+  constructor(
+    @InjectRepository(FavMoviesEntity)
+    private favMoviesRepository: Repository<FavMoviesEntity>,
+    private userService: UserService,
+    private http: HttpService,
+  ) {}
 
   getAllMovies(
     genresIds: number[],
@@ -33,5 +47,67 @@ export class FavMoviesService {
         `${this.apiUrl}/movie/${id}?api_key=${process.env.API_KEY}&language=en-US`,
       )
       .pipe(map((res) => res.data));
+  }
+
+  async addFavMovies(
+    id: number,
+    { login }: UserEntity,
+  ): Promise<FavMoviesEntity> {
+    const user = await this.userService.getUserByLogin(login);
+
+    const favMoviesEntity: FavMoviesEntity = new FavMoviesEntity();
+    favMoviesEntity.moviesId = id;
+    favMoviesEntity.watched = false;
+    favMoviesEntity.user = user;
+
+    try {
+      await favMoviesEntity.save();
+      return favMoviesEntity;
+    } catch (e) {
+      throw new InternalServerErrorException();
+    }
+  }
+
+  async setFavMovieAsWatched(moviesId: number): Promise<FavMoviesEntity> {
+    const movie = await this.favMoviesRepository.findOne({ moviesId });
+    movie.watched = true;
+
+    try {
+      await movie.save();
+      return movie;
+    } catch (e) {
+      throw new InternalServerErrorException();
+    }
+  }
+
+  async removeFavMovies(moviesId: number): Promise<FavMoviesEntity> {
+    const movie = await this.favMoviesRepository.findOne({ moviesId });
+
+    if (!movie) {
+      throw new BadRequestException('Movie not found');
+    }
+
+    await this.favMoviesRepository.delete(movie.id);
+
+    return movie;
+  }
+
+  async getFavMovies({
+    login,
+  }: UserEntity): Promise<AxiosResponse<FavMoviesDto, any>[]> {
+    const user = await this.userService.getUserByLogin(login);
+
+    const movies = await this.favMoviesRepository.find({
+      where: { user: user.id },
+    });
+    const result = await Promise.all(
+      movies.map((movie) => {
+        let moviesDetails = this.getMoviesDetails(movie.moviesId);
+        let movieIdData = lastValueFrom(moviesDetails);
+        return movieIdData;
+      }),
+    );
+
+    return result;
   }
 }
